battle1: four more -- the two sound-pan helpers, the particle scatter, the scratchpad descriptor
func_800BBA84, func_800C2FD4, func_800C57B0, func_800BA360. All MATCH,
`make build` 19x OK with every battle object byte-identical to expected/.

Two findings, both recorded in CLAUDE.md:

* `x < 0` compiles to `srl x,31` through an `s32` local and to `slti x,0`
  through a `u8` one -- `do_store_flag`'s sign-test shortcut only fires when
  the destination is the comparison's own mode. One row, one word, and every
  respelling of the comparison itself (`0 > v`, the ternary, `* 8`, `s16`,
  `u32`) is the `srl` form. func_800BA360 turned on exactly that.

* A `u16` parameter gets no entry mask when a prototype is in scope, so the
  standing "a plain `move` cannot come from a narrow parameter" rule does not
  apply to this codebase's prototyped calls -- and the parameter's width has
  to be read off the *callers*, where `lhu` means `u16` and `lh` means `s32`.
  Reading func_800BBA84 as `(s32, s16, s32)` matched the function under test
  and silently moved two loads in battle2.c; `(u16, s16, s32)` matches both.

func_800C57B0 also needed a named `base` local: without it gcc folds the
+0x3800 into the symbol and derives the record cursor from the byte array,
where the target materialises one base and derives both from it.

+++ src/battle/battle_private.h
@@ -603,7 +603,7 @@ static void func_800BA40C();
 static void func_800BB030(s16 arg0);
 void func_800BB2A8(u8);
 void func_800BB9B8(s32);
-void func_800BBA84(u16 arg0, s32 arg1, s32 arg2);
+void func_800BBA84(u16 arg0, s16 arg1, s32 arg2);
 static void func_800C0480(s16);
 static void func_800C0630(s16);
 static void func_800C0970(s16);
